refactor: move project membership check from access to project
